rework files and namespaces
1. split into .h/.cpp files
2. separated main types and aditional ones into namespaces
3. added documentation into header
